Temporary rework
I might actually be the dumbest person on the planet. I spent an
unbelievable amount of time trying to figure out the correct formula for
indexing when splitting ops and I could for the life of me not find the
correct answer no matter how hard I though about it and how sure I was
the formula *had* to be correct. I actually wanted to just quit
programming or at the very least this project because of how frustrating
this was. I did not consider for a second for a second that the
variables used in the formula could have the wrong values. Up until 30
minutes ago. You can figure out the rest.
UGHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHH