Fix `rng=1737055237313153` temporarily
The issue was that in the previous one only equal ops where checked for
overlapping. This was done to prevent checking overlaps for non same
size buffers, which is not yet implemented, but that could lead to an
issue where ops of different types where still operating on the same
values.