Fix non linear offsets (terrible but fixed)
Needs to be reworked already. This basically was an issue with the way
loops were recognized and dim_info stores the index information.
Essentially dim_info only does linear transformation that loop back
around after a while, which can not not describe some configuration of
offsets that can occur ever from fairly natural sources. Note that we
are not using tables for the indices because that artificially limits
the size of the assignment loops as the OpenCL kernels only have so much
local memory and using global memory seems hilariously slow.